refactor(style): move components styles to css files
Related work item: GH-3
