+++ src/components/body/Body.css
@@ -1,4 +1,3 @@
-export const style = `
 :host {
   width: 100%;
   max-height: 100%;
@@ -10,9 +9,9 @@ main {
   box-sizing: border-box;
 }
 :host::-webkit-scrollbar-track {
-    -webkit-box-shadow: inset 0 0 0.375rem rgba(0, 0, 0, 0.3);
-    background-color: transparent;
-  }
+  -webkit-box-shadow: inset 0 0 0.375rem rgba(0, 0, 0, 0.3);
+  background-color: transparent;
+}
 
 :host::-webkit-scrollbar {
   width: 0.75rem;
@@ -22,4 +21,3 @@ main {
 :host::-webkit-scrollbar-thumb {
   background-color: var(--brand);
 }
-`;
