+++ src/components/events/Events.css
@@ -1,4 +1,3 @@
-export const style = `
 section {
   width: 100%;
 }
@@ -40,4 +39,3 @@ li {
     justify-content: flex-start;
   }
 }
-`;
