+++ src/components/events/SingleEvent/SingleEvent.css
@@ -1,4 +1,3 @@
-export const style = `
 a {
   position: relative;
   text-decoration: none;
@@ -66,4 +65,3 @@ a {
     -webkit-line-clamp: 8;
   }
 }
-`;
