+++ src/components/footer/Footer.css
@@ -1,4 +1,3 @@
-export const style = `
 :host {
   width: 100%;
 }
@@ -28,4 +27,3 @@ a {
 a:hover {
   opacity: 100%;
 }
-`;
