+++ src/components/heaedr/Header.css
@@ -1,4 +1,3 @@
-export const style = `
 kk-heaedr {
   width: 100%;
 }
@@ -57,4 +56,3 @@ header nav kk-tabs {
     padding-left: 67%;
   }
 }
-`;
