+++ src/components/icon/Icon.css
@@ -1,4 +1,3 @@
-export const style = `
 :host {
   display: flex;
   justify-content: center;
@@ -17,4 +16,3 @@ svg:hover, svg[aria-selected="true"] {
   filter: drop-shadow(var(--shadow-around-level-0) var(--brand));
   fill: var(--brand);
 }
-`;
