+++ src/components/labeledAvatar/LabeledAvatar.css
@@ -1,4 +1,3 @@
-export const style = `
 :host {
   display: block;
 }
@@ -26,4 +25,4 @@ export const style = `
   font-size: var(--font-size-s);
   max-width: 25ch;
   text-align: center;
-`;
+}
