+++ src/components/labeledUrlAvatar/LabeledUrlAvatar.css
@@ -1,4 +1,3 @@
-export const style = `
 :host {
   width: 100%;
 }
@@ -31,4 +30,3 @@ a {
 .section-title {
   padding-left: var(--font-size-l);
 }
-`;
