+++ src/components/projects/Projects.css
@@ -1,4 +1,3 @@
-export const style = `
 ul {
   margin: 0;
   padding: 0;
@@ -12,4 +11,3 @@ ul {
 .project {
   padding: var(--spacing-xl) 0;
 }
-`;
