+++ src/components/projects/SingleProject/SingleProject.css
@@ -1,4 +1,3 @@
-export const style = `
 :host {
   width: 100%;
 }
@@ -41,12 +40,19 @@ details[open] summary,
 .project__picture__description {
   opacity: 100%;
 }
+
 .project__info {
   padding-left: var(--spacing-l);
 }
-.project__picture__description {
-  text-align: center;
-  font-weigh: bold;
+.project__creators {
+  display: flex;
+  flex-wrap: wrap;
+  align-items: center;
+  justify-content: space-around;
+}
+.creator {
+  cursor: pointer;
+  padding: var(--spacing-m);
 }
 .picture-wrapper {
   display: flex;
@@ -54,22 +60,23 @@ details[open] summary,
   align-items: center;
   justify-content: center;
 }
+.project__picture__description {
+  text-align: center;
+  font-weight: bold;
+}
 .project__picture {
   max-width: 100%;
-  padding-bottom: var(--spacing-m);
 }
-
 @media (min-width: 1350px) {
-  .project__info {
+  .project__content--wrapper {
     display: grid;
     grid-template-columns: 1fr 1fr;
     align-items: start;
   }
-  .project__additional-info {
+  .project__demo {
     height: 100%;
     display: flex;
     flex-direction: column;
     justify-content: space-around;
   }
 }
-`;
