+++ src/components/socialMedia/SocialMedia.css
@@ -1,4 +1,3 @@
-export const style = `
 section {
   display: flex;
   justify-content: center;
@@ -8,4 +7,3 @@ a {
   font-size: 0;
   padding: var(--spacing-m);
 }
-`;
