+++ src/components/tabs/SingleTab/SingleTab.css
@@ -1,4 +1,3 @@
-export const style = `
 button {
   user-select: text;
   outline: none;
@@ -239,4 +238,3 @@ button {
     background-position: bottom;
   }
 }
-`;
