+++ src/components/tabs/TabIndicator/TabIndicator.css
@@ -1,4 +1,3 @@
-export const style = `
 .indicator {
   position: relative;
   display: block;
@@ -8,4 +7,3 @@ export const style = `
   filter: drop-shadow(var(--shadow-around-level-1) var(--brand));
   transition: left var(--long-duration) var(--ease-in-out-quint);
 }
-`;
