+++ src/components/tabs/Tabs.css
@@ -1,4 +1,3 @@
-export const style = `
 .tabs {
   position: relative;
   list-style: none;
@@ -19,4 +18,3 @@ export const style = `
 .tabs li {
   flex: 1;
 }
-`;
