+++ src/components/teams/SingleTeam/SingleTeam.css
@@ -1,4 +1,3 @@
-export const style = `
 :host {
   width: 100%;
 }
@@ -51,7 +50,7 @@ h2 {
 }
 .single-section {
   padding-left: var(--spacing-xl);
-} 
+}
 .section-description {
   padding-left: var(--spacing-m);
 }
@@ -69,4 +68,3 @@ h2 {
     -webkit-line-clamp: 8;
   }
 }
-`;
