+++ src/components/teams/Teams.css
@@ -1,4 +1,3 @@
-export const style = `
 ul {
   margin: 0;
   padding: 0;
@@ -7,4 +6,3 @@ ul {
 .team {
   padding: var(--spacing-xl) 0;
 }
-`;
