+++ src/components/urlIcon/UrlIcon.css
@@ -1,8 +1,6 @@
-export const style = `
 :host {
   width: 100%;
 }
 a {
   font-size: 0;
 }
-`;
